+++ commons/src/main/java/commons/ScoreMessage.java
@@ -14,10 +14,10 @@ public class ScoreMessage extends Message<Integer> {
     private final UUID id;
 
     @JsonProperty("answer")
-    private final int answer;
+    private final long answer;
 
     @JsonProperty("option")
-    private final int option;
+    private final long option;
 
 
     //for open questions
@@ -25,8 +25,8 @@ public class ScoreMessage extends Message<Integer> {
     public ScoreMessage(final @JsonProperty("nick") String nick,
                         final @JsonProperty("time") int time,
                         final @JsonProperty("type") String type,
-                        final @JsonProperty("answer") int answer,
-                        final @JsonProperty("option") int option,
+                        final @JsonProperty("answer") long answer,
+                        final @JsonProperty("option") long option,
                         final @JsonProperty("id") UUID id) {
         super(nick, time);
         this.type = type;
@@ -43,11 +43,11 @@ public class ScoreMessage extends Message<Integer> {
         return id;
     }
 
-    public int getAnswer() {
+    public long getAnswer() {
         return answer;
     }
 
-    public int getOption() {
+    public long getOption() {
         return option;
     }
